Add project deferent modules
