+++ src/index.html
@@ -22,7 +22,7 @@
     <section class="todo-container" aria-labelledby="todo-header">
       <header class="todo-header">
         <h1 id="todo-header" class="primary-header">Todo</h1>
-        <img class="themeChanger" width="25" height="25" src="../src/images/icon-moon.svg" alt="change theme icon">
+        <img class="themeChanger" width="25" height="25" src="#" alt="change theme icon">
       </header>
       <div class="todo-body">
         <div class="bg-todo px-4 rounded mb-25 mt-50">
@@ -35,32 +35,32 @@
             <li class="todo-item px-4">
               <div class="check active"></div>
               <p class="task active">Complete on;ine javascript course</p>
-              <img class="cross-icon" src="./images/icon-cross.svg" width="15" height="15" alt="cross icon">
+              <img class="cross-icon" src="#" width="15" height="15" alt="cross icon">
             </li>
             <li class="todo-item px-4">
               <div class="check"></div>
               <p class="task">Jog around the park 3x</p>
-              <img class="cross-icon" src="./images/icon-cross.svg" width="15" height="15" alt="cross icon">
+              <img class="cross-icon" src="#" width="15" height="15" alt="cross icon">
             </li>
             <li class="todo-item px-4">
               <div class="check"></div>
               <p class="task">10 minutes meditation</p>
-              <img class="cross-icon" src="./images/icon-cross.svg" width="15" height="15" alt="cross icon">
+              <img class="cross-icon" src="#" width="15" height="15" alt="cross icon">
             </li>
             <li class="todo-item px-4">
               <div class="check"></div>
               <p class="task">Read for One hour</p>
-              <img class="cross-icon" src="./images/icon-cross.svg" width="15" height="15" alt="cross icon">
+              <img class="cross-icon" src="#" width="15" height="15" alt="cross icon">
             </li>
             <li class="todo-item px-4">
               <div class="check"></div>
               <p class="task">Pick up groceries</p>
-              <img class="cross-icon" src="./images/icon-cross.svg" width="15" height="15" alt="cross icon">
+              <img class="cross-icon" src="#" width="15" height="15" alt="cross icon">
             </li>
             <li class="todo-item px-4">
               <div class="check"></div>
               <p class="task">Complete todo App project</p>
-              <img class="cross-icon" src="./images/icon-cross.svg" width="15" height="15" alt="cross icon">
+              <img class="cross-icon" src="#" width="15" height="15" alt="cross icon">
             </li>
           </ul>
           <div class="state px-4">
