+++ src/modules/mainEvents.js
@@ -1,0 +1,21 @@
+// project HTML page images
+import cross from '../images/icon-cross.svg';
+import moon from '../images/icon-moon.svg';
+import sun from '../images/icon-sun.svg';
+
+export default function mainEvent () {
+  const crossIcon = document.querySelectorAll('.cross-icon');
+  document.querySelector('.themeChanger').src = moon;
+  crossIcon.forEach((element) => {
+    element.src = cross;
+  });
+
+  const todoinput = document.querySelector('.todo-input');
+  todoinput.addEventListener('focus', (e) => {
+    e.target.placeholder = 'Currently typing...';
+  });
+
+  todoinput.addEventListener('blur', (e) => {
+    e.target.placeholder = 'Create a new todo...';
+  });
+};
